adopters: remove dbus-objects
Unfortunately, after a few users raised concern about the enforcability
of this license I had to remove it.

I think the main concern is about the legalese included in the license
and users not feeling comfortable enough to use it. Point 3 of
'Compliance with Human Rights Principles and Human Rights Laws.' comes
to mind. For eg, from my interpretation the Licensee shouldn't be
forced to indemnify the Licensor if someone makes a violation claim
agaisnt them, just in the case they are found to be in violation. But
this is not nessarily clear. I even think there is a possibility for my
interpretation to be incorrect.

Signed-off-by: Filipe Laíns <lains@riseup.net>